Fix various compiler warnings.

+++ src/Packages/Teem/Dataflow/Modules/Unu/UnuFlip.cc
@@ -107,7 +107,7 @@ UnuFlip::execute()
   out->copy_properties(nrrd_handle.get_rep());
 
   // Copy the axis kinds
-  for (int i=0; i<nin->dim; i++) {
+  for (size_t i=0; i<nin->dim; i++) {
     nout->axis[i].kind = nin->axis[i].kind;
   }
